+++ express-app/app.js
@@ -3,10 +3,8 @@ const express = require('express');
 const path = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
-const mysql = require('mysql');
 
 const indexRouter = require('./routes/index');
-const usersRouter = require('./routes/users');
 
 const app = express();
 
@@ -21,20 +19,6 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
-app.use('/users', usersRouter);
-
-// MySQLへの接続
-const con = mysql.createConnection({
-  host: 'db',
-  user: 'root',
-  password: 'password'
-});
-
-con.connect(function(err) {
-  if (err) throw err;
-  console.log('Connected');
-  database: 'sample_db'
-});
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
