+++ express-app/routes/index.js
@@ -1,9 +1,27 @@
 const express = require('express');
 const router = express.Router();
+const mysql = require('mysql');
+
+ // MySQLへの接続
+const con = mysql.createConnection({
+host: 'db',
+user: 'root',
+password: 'password',
+database: 'sample_db'
+});
+
+con.connect(function(err) {
+	if (err) throw err;
+	console.log('Connected');
+});
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-  res.render('index', { title: 'Express on docker' });
+  const sql = "select * from sampledata"
+  con.query(sql, function (err, result, fields) {
+    if (err) throw err;
+      res.send(result);
+  });
 });
 
 module.exports = router;
